% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dna_atac_set.R
\name{buildDNASequences}
\alias{buildDNASequences}
\title{Function that extracts the DNA Sequences from the region's data obtained in the previous step,
using the BSGenome library.}
\usage{
buildDNASequences(region_set, region_type, length_cutoff = Inf)
}
\arguments{
\item{region_set}{Set that contains the region information from the specified region type}

\item{region_type}{String with the type of the region from which to extract the DNA sequences}

\item{length_cutoff}{Integer with the minimum length required for a sequence to be selected for training}
}
\value{
A set with the DNA sequence of the specified region type
}
\description{
Function that extracts the DNA Sequences from the region's data obtained in the previous step,
using the BSGenome library.
}
