% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dna_atac_set.R
\name{buildRegionSets}
\alias{buildRegionSets}
\title{Main function that builds the training set for each type of region (i.e. promoters, 3' UTR and 5' UTR),
calling the rest of the functions of the present script}
\usage{
buildRegionSets(path_to_gtf_file, path_to_coverage_matrix, output_dir)
}
\arguments{
\item{path_to_gtf_file}{String with the path to the most recent mouse GTF file}

\item{path_to_coverage_matrix}{String with the path where the coverage matrix will be saved}

\item{output_dir}{String with the output folder where the final sets will be stored}
}
\value{

}
\description{
Main function that builds the training set for each type of region (i.e. promoters, 3' UTR and 5' UTR),
calling the rest of the functions of the present script
}
