% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dna_atac_set.R
\name{computeCoverageMatrix}
\alias{computeCoverageMatrix}
\title{Function that, given two BAM files from the mouse genome (before and one hour after the injection),
builds the mean coverage matrix for each protein coding gene}
\usage{
computeCoverageMatrix(
  path_to_gtf_file,
  path_to_bam_data,
  path_to_coverage_matrix
)
}
\arguments{
\item{path_to_gtf_file}{String with the path to the most recent mouse GTF file}

\item{path_to_bam_data}{String with the folder that contains the BAM files}

\item{path_to_coverage_matrix}{String with the path where the mean coverage matrix will be saved}
}
\value{
The mean coverage matrix obtained from the .BAM data
}
\description{
Function that, given two BAM files from the mouse genome (before and one hour after the injection),
builds the mean coverage matrix for each protein coding gene
}
