% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dna_atac_set.R
\name{createFinalDataFrame}
\alias{createFinalDataFrame}
\title{Final function the builds the training dataframe using the information obtained in the previous functions}
\usage{
createFinalDataFrame(
  region_set,
  dna_sequences,
  fold_changes,
  df_mean_coverage,
  output_path
)
}
\arguments{
\item{region_set}{Set that contains the region information from the region type of the current iteration}

\item{dna_sequences}{DNA sequences from the region type of the current iteration}

\item{fold_changes}{Vector with the fold changes in the ATAC-Seq, which will be the output variable of our model}

\item{df_mean_coverage}{Dataframe with the mean coverage matrix built in the first step of the script}

\item{output_path}{String with the output path in which to save the dataframe}
}
\value{

}
\description{
Final function the builds the training dataframe using the information obtained in the previous functions
}
