% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dna_atac_set.R
\name{getCoverage}
\alias{getCoverage}
\title{Function that, given the characteristics of a gene (chromosome, start position and end position),
calculates the coverage of each base pair of the gene and applies a normalized function to that coverage.}
\usage{
getCoverage(bamPath, seqid, start, end, FUN, bamLength)
}
\arguments{
\item{bamPath}{String with the path to the BAM file to use to extract the coverage}

\item{seqid}{ID of the chromosome of the present gene}

\item{start}{Start position of the gene in the mouse genome}

\item{end}{End position of the gene in the mouse genome}

\item{FUN}{Function to be applied to the coverage}

\item{bamLength}{Length of the BAM file used to normalize}
}
\value{
The mean normalized coverage for the specified gene
}
\description{
Function that, given the characteristics of a gene (chromosome, start position and end position),
calculates the coverage of each base pair of the gene and applies a normalized function to that coverage.
}
