% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dna_atac_set.R
\name{loadRegionData}
\alias{loadRegionData}
\title{Function that loads the region's data to use in each iteration (i.e. promoters, 3' UTR or 5' UTR)}
\usage{
loadRegionData(path_to_gtf_file, region_type)
}
\arguments{
\item{path_to_gtf_file}{String with the path to the most recent mouse GTF file}

\item{region_type}{String with the type of region from which to extract the data}
}
\value{
A dataframe with the main data of the specified region
}
\description{
Function that loads the region's data to use in each iteration (i.e. promoters, 3' UTR or 5' UTR)
}
